refactor: URL構造変更のため、area / shop のプレフィックスを削除
