+++ frontend/lib/fetchShop.ts
@@ -1,6 +1,6 @@
 export const fetchShop = async (areaRomaji: string, subAreaRomaji: string, shopId: string) => {
   const res = await fetch(
-    `${process.env.NEXT_PUBLIC_API_BASE_URL}/shop/${areaRomaji}/${subAreaRomaji}/${shopId}`,
+    `${process.env.NEXT_PUBLIC_API_BASE_URL}/${areaRomaji}/${subAreaRomaji}/${shopId}`,
     { cache: "force-cache" }
   );
